Inner classes of TwoPhaseQueryUnfolder moved to upper level.
