+++ core/model/src/main/java/it/unibz/inf/ontop/injection/CoreSingletons.java
@@ -3,6 +3,7 @@ package it.unibz.inf.ontop.injection;
 import it.unibz.inf.ontop.constraints.HomomorphismFactory;
 import it.unibz.inf.ontop.dbschema.DatabaseInfoSupplier;
 import it.unibz.inf.ontop.iq.node.normalization.ConstructionSubstitutionNormalizer;
+import it.unibz.inf.ontop.iq.tools.UnionBasedQueryMerger;
 import it.unibz.inf.ontop.iq.type.NotYetTypedBinaryMathOperationTransformer;
 import it.unibz.inf.ontop.iq.type.NotYetTypedEqualityTransformer;
 import it.unibz.inf.ontop.iq.type.PartiallyTypedSimpleCastTransformer;
@@ -57,5 +58,7 @@ public interface CoreSingletons {
     PartiallyTypedSimpleCastTransformer getPartiallyTypeSimpleCastTransformer();
 
     DatabaseInfoSupplier getDatabaseInfoSupplier();
+
+    UnionBasedQueryMerger getUnionBasedQueryMerger();
     // TODO: complete
 }
