+++ core/model/src/main/java/it/unibz/inf/ontop/injection/impl/CoreSingletonsImpl.java
@@ -9,6 +9,7 @@ import it.unibz.inf.ontop.injection.IntermediateQueryFactory;
 import it.unibz.inf.ontop.injection.OntopModelSettings;
 import it.unibz.inf.ontop.injection.QueryTransformerFactory;
 import it.unibz.inf.ontop.iq.node.normalization.ConstructionSubstitutionNormalizer;
+import it.unibz.inf.ontop.iq.tools.UnionBasedQueryMerger;
 import it.unibz.inf.ontop.iq.type.NotYetTypedBinaryMathOperationTransformer;
 import it.unibz.inf.ontop.iq.type.NotYetTypedEqualityTransformer;
 import it.unibz.inf.ontop.iq.type.PartiallyTypedSimpleCastTransformer;
@@ -41,6 +42,7 @@ public class CoreSingletonsImpl implements CoreSingletons {
     private final NotYetTypedBinaryMathOperationTransformer notYetTypedBinaryMathOperationTransformer;
     private final PartiallyTypedSimpleCastTransformer partiallyTypedSimpleCastTransformer;
     private final DatabaseInfoSupplier databaseInfoSupplier;
+    private final UnionBasedQueryMerger unionBasedQueryMerger;
 
     @Inject
     private CoreSingletonsImpl(TermFactory termFactory, TypeFactory typeFactory,
@@ -56,7 +58,7 @@ public class CoreSingletonsImpl implements CoreSingletons {
                                NotYetTypedEqualityTransformer notYetTypedEqualityTransformer,
                                NotYetTypedBinaryMathOperationTransformer notYetTypedBinaryMathOperationTransformer,
                                PartiallyTypedSimpleCastTransformer partiallyTypedSimpleCastTransformer,
-                               DatabaseInfoSupplier databaseInfoSupplier) {
+                               DatabaseInfoSupplier databaseInfoSupplier, UnionBasedQueryMerger unionBasedQueryMerger) {
         this.termFactory = termFactory;
         this.typeFactory = typeFactory;
         this.functionSymbolFactory = functionSymbolFactory;
@@ -74,6 +76,7 @@ public class CoreSingletonsImpl implements CoreSingletons {
         this.notYetTypedBinaryMathOperationTransformer = notYetTypedBinaryMathOperationTransformer;
         this.partiallyTypedSimpleCastTransformer = partiallyTypedSimpleCastTransformer;
         this.databaseInfoSupplier = databaseInfoSupplier;
+        this.unionBasedQueryMerger = unionBasedQueryMerger;
     }
 
     @Override
@@ -159,5 +162,10 @@ public class CoreSingletonsImpl implements CoreSingletons {
     @Override
     public DatabaseInfoSupplier getDatabaseInfoSupplier() {
         return databaseInfoSupplier;
+    }
+
+    @Override
+    public UnionBasedQueryMerger getUnionBasedQueryMerger() {
+        return unionBasedQueryMerger;
     }
 }
